64-bit transition part 6: remainer of additaments to libast (#955)
This is the sixth of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The reminder of SFIO located in the disc folder.
- The libast tm and tmx sublibraries.
- The rest of the libast path sublibrary.
- The rest of the libast headers and the cdt(3) man page.
  - ast.mb.cur_max is uint32_t to save 8 bytes (64 bytes -> 56
    bytes total for struct size). MB_CUR_MAX is actually a 64-bit
    size_t, but I doubt it'll ever actually exceed the 64-bit
    limit.
- The vmalloc wrapper's header file.
- The AST dirlib.
- Minor additional warning fix in optget() (re: ed1c3448).

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 1,947 => 1,580 => 37 (progression from
part 5 => part 6 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/comp/setlocale.c
@@ -2210,7 +2210,7 @@ set_ctype(Lc_category_t* cp)
 		ast.mb.conv = utf8_wctomb;
 		ast.mb.alpha = utf8_alpha;
 	}
-	else if ((locales[cp->internal]->flags & LC_default) || (ast.mb.cur_max = MB_CUR_MAX) <= 1 || !(ast.mb.len = mblen) || !(ast.mb.towc = mbtowc))
+	else if ((locales[cp->internal]->flags & LC_default) || (ast.mb.cur_max = (uint32_t)MB_CUR_MAX) <= 1 || !(ast.mb.len = mblen) || !(ast.mb.towc = mbtowc))
 	{
 		ast.mb.cur_max = 1;
 		ast.mb.len = 0;
